Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -2,7 +2,7 @@
 
 
 int main() {
-    int n;
+    float n;
     scanf("%f",&n);
     printf("You entered: ",n);
     return 0;
